update service Rent

+++ API/digital-booking-c3-e-1/src/main/java/com/digitalbooking/digitalbooking/domain/rent/service/RentService.java
@@ -26,32 +26,41 @@ public class RentService {
     @Autowired
     RepositoryUser repositoryUser;
 
-    public String deleteRent(Rent rent){
+    public String deleteRent(Rent rent, String userEmail){
+        RentDTO rentDTO = rentRepository.findByIdAndState(rent.getId()).orElseThrow(()->new ExceptionNullValue("El alquiler que intenta actualizar no se encontró"));
+        validateRentOwnership(rentDTO.getUser().getId(), userEmail);
         rentRepository.deleteRent(rent.getId());
-        return "Alquiler eliminado correctamente";
+        return "Alquiler cancelado correctamente";
     }
 
     public Long createRent(Rent rent, String userEmail){
         repositoryProduct.findByIdAndIsDelete(rent.getProduct().getId()).orElseThrow(()->new ExceptionNullValue("Producto no encontrado"));
-        UserDTO user = repositoryUser.findByEmail(userEmail).orElseThrow(()->new ExceptionNullValue("Usuario no encontrado"));
-        if(!Objects.equals(user.getId(), rent.getUser().getId())){
-            throw new ExceptionInvalidValue("El id del usuario no pertenece al suyo, no puede crear alquileres para otros usuarios");
-        }
+        validateRentOwnership(rent.getUser().getId(), userEmail);
         return rentRepository.createRent(rent);
     }
 
-    public String updateRent(Rent rent) throws Exception {
-        rentRepository.findByIdAndState(rent.getId()).orElseThrow(()->new ExceptionNullValue("Alquiler no encontrado"));
+    public String updateRent(Rent rent, String userEmail) throws Exception {
+        RentDTO rentDto = rentRepository.findByIdAndState(rent.getId()).orElseThrow(()->new ExceptionNullValue("Alquiler no encontrado"));
+        validateRentOwnership(rentDto.getUser().getId(), userEmail);
         rentRepository.updateRent(rent);
         return "Alquiler actualizado correctamente";
     }
 
-    public List<RentDTO> getRents(){
-        return rentRepository.getAll();
+    public List<RentDTO> getRents(String userEmail){
+        UserDTO user = repositoryUser.findByEmail(userEmail).orElseThrow(()->new ExceptionNullValue("Usuario no encontrado"));
+        return rentRepository.getAll(user.getId());
     }
 
-    public RentDTO getRent(Long id) {
-        return rentRepository.findByIdAndState(id).orElseThrow(()->new ExceptionNullValue("Alquiler no encontrado"));
+    public RentDTO getRent(Long id, String userEmail) {
+        RentDTO rentDto = rentRepository.findByIdAndState(id).orElseThrow(()->new ExceptionNullValue("Alquiler no encontrado"));
+        validateRentOwnership(rentDto.getUser().getId(), userEmail);
+        return rentDto;
     }
 
+    private void validateRentOwnership(Long idOwner, String userEmail) {
+        UserDTO user = repositoryUser.findByEmail(userEmail).orElseThrow(()->new ExceptionNullValue("Usuario no encontrado"));
+        if(!Objects.equals(user.getId(), idOwner)){
+            throw new ExceptionInvalidValue("El id del usuario no te pertenece, no puedes crear o acceder a alquileres de otros usuarios");
+        }
+    }
 }
